Use `np.int32` instead of `np.int64` throughout to comply with C default `int` and possibly save some memory.

+++ motep/loss.py
@@ -110,7 +110,7 @@ class LossFunctionForces:
 
     Attributes
     ----------
-    idcs_frc : npt.NDArray[np.int64]
+    idcs_frc : npt.NDArray[np.int32]
         Indices of images that have forces.
 
     """
@@ -201,7 +201,7 @@ class LossFunctionStress:
 
     Attributes
     ----------
-    idcs_str : npt.NDArray[np.int64]
+    idcs_str : npt.NDArray[np.int32]
         Indices of images that have 3D cells.
 
     """
